requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
groq==0.4.1
python-dotenv==1.0.0
reportlab==4.0.4
beautifulsoup4==4.12.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Flask-WTF==1.1.1
WTForms==3.0.1
beautifulsoup4
requests